# gpt-actions.openapi.yaml
openapi: 3.1.0
info:
  title: Windows LocalControl GPT Actions
  description: Private Custom GPT action schema for a single-user Windows LocalControl
    bridge. Full-control execution mode is enabled; use /execution/logs to retrieve
    command and terminal output.
  version: 0.1.0
paths:
...
  /artifacts/create_text:
    post:
      tags:
      - artifacts
      summary: Artifacts Create Text
      operationId: artifacts_create_text
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtifactCreateTextRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /artifacts/fetch_url:
    post:
      tags:
      - artifacts
      summary: Artifacts Fetch Url
      operationId: artifacts_fetch_url
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtifactFetchUrlRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /artifacts/from_path:
    post:
      tags:
      - artifacts
      summary: Artifacts From Path
      operationId: artifacts_from_path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtifactFromPathRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /artifacts/list:
    post:
      tags:
      - artifacts
      summary: Artifacts List
      operationId: artifacts_list
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtifactListRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /artifacts/{artifact_id}:
    get:
      tags:
      - artifacts
      summary: Artifacts Get
      operationId: artifacts_get
      security:
      - HTTPBearer: []
      parameters:
      - name: artifact_id
        in: path
        required: true
        schema:
          type: string
          title: Artifact Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactInfo'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /artifacts/{artifact_id}/write_to_path:
    post:
      tags:
      - artifacts
      summary: Artifacts Write To Path
      operationId: artifacts_write_to_path
      security:
      - HTTPBearer: []
      parameters:
      - name: artifact_id
        in: path
        required: true
        schema:
          type: string
          title: Artifact Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtifactWriteToPathRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactInfo'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ArtifactCreateTextRequest:
      properties:
        name:
          type: string
          title: Name
        content:
          type: string
          title: Content
        mime_type:
          type: string
          title: Mime Type
          default: text/plain
        encoding:
          type: string
          title: Encoding
          default: utf-8
      additionalProperties: false
      type: object
      required:
      - name
      - content
      title: ArtifactCreateTextRequest
...
    ArtifactFetchUrlRequest:
      properties:
        url:
          type: string
          title: Url
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      additionalProperties: false
      type: object
      required:
      - url
      title: ArtifactFetchUrlRequest
    ArtifactFromPathRequest:
      properties:
        path:
          type: string
          title: Path
        copy:
          type: boolean
          title: Copy
          default: false
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      additionalProperties: false
      type: object
      required:
      - path
      title: ArtifactFromPathRequest
    ArtifactInfo:
      properties:
        artifact_id:
          type: string
          title: Artifact Id
        name:
          type: string
          title: Name
        size:
          type: integer
          title: Size
        mime_type:
          type: string
          title: Mime Type
        sha256:
          type: string
          title: Sha256
        created_at:
          type: string
          title: Created At
        source:
          type: string
          title: Source
        local_path:
          type: string
          title: Local Path
        managed:
          type: boolean
          title: Managed
      additionalProperties: false
      type: object
      required:
      - artifact_id
      - name
      - size
      - mime_type
      - sha256
      - created_at
      - source
      - local_path
      - managed
      title: ArtifactInfo
    ArtifactListRequest:
      properties:
        max_results:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Max Results
          default: 100
      additionalProperties: false
      type: object
      title: ArtifactListRequest
    ArtifactListResponse:
      properties:
        artifacts:
          items:
            $ref: '#/components/schemas/ArtifactInfo'
          type: array
          title: Artifacts
      additionalProperties: false
      type: object
      required:
      - artifacts
      title: ArtifactListResponse
...
    ArtifactWriteToPathRequest:
      properties:
        path:
          type: string
          title: Path
        overwrite:
          type: boolean
          title: Overwrite
          default: false
        create_parents:
          type: boolean
          title: Create Parents
          default: true
      additionalProperties: false
      type: object
      required:
      - path
      title: ArtifactWriteToPathRequest
...
    ErrorResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: false
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
      additionalProperties: false
      type: object
      required:
      - code
      - message
      title: ErrorResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://YOUR-RESERVED-NGROK-DOMAIN.ngrok-free.app